refactor: 딜러와 Account 공유를 하기 때문에 Account 예외처리 기능을 Players로 이전
